chore: change structure to its onw file for Stripe checkout session and user role management in Firestore
